# api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/identity/bootstrap": {
      "get": {
        "operationId": "IdentityController_bootstrap",
        "parameters": [
          {
            "name": "agentId",
            "required": false,
            "in": "query",
            "description": "Agent ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Bootstrap identity data for agent/user pair",
        "tags": [
          "identity"
        ]
      }
    },
...
    "/v1/identity/agents/{id}": {
      "get": {
        "operationId": "IdentityController_getAgent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agent ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a single agent with capability profiles and trust summary",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/contracts": {
      "post": {
        "operationId": "IdentityController_createContract",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDelegationContractDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a delegation contract",
        "tags": [
          "identity"
        ]
      },
      "get": {
        "operationId": "IdentityController_listContracts",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agentId",
            "required": false,
            "in": "query",
            "description": "Filter by delegated agent ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isTemplate",
            "required": false,
            "in": "query",
            "description": "Filter by template flag",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List delegation contracts",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/contracts/{id}": {
      "get": {
        "operationId": "IdentityController_getContract",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contract ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a delegation contract by ID",
        "tags": [
          "identity"
        ]
      },
      "put": {
        "operationId": "IdentityController_updateContract",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contract ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDelegationContractDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a delegation contract",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/contracts/{id}/complete": {
      "post": {
        "operationId": "IdentityController_completeContract",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contract ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteContractRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Complete or fail a delegation contract",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/challenges": {
      "post": {
        "operationId": "IdentityController_createChallenge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChallengeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a challenge",
        "tags": [
          "identity"
        ]
      },
      "get": {
        "operationId": "IdentityController_listChallenges",
        "parameters": [
          {
            "name": "contractId",
            "required": false,
            "in": "query",
            "description": "Filter by contract ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by resolution status (open/resolved/dismissed)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by challenge type",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List challenges",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/challenges/{id}": {
      "get": {
        "operationId": "IdentityController_getChallenge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Challenge ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a challenge by ID",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/challenges/{id}/resolve": {
      "post": {
        "operationId": "IdentityController_resolveChallenge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Challenge ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveChallengeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Resolve a challenge",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/teams": {
      "get": {
        "operationId": "IdentityController_listTeams",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all teams",
        "tags": [
          "identity"
        ]
      },
      "post": {
        "operationId": "IdentityController_createTeam",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a team profile",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/teams/{id}": {
      "get": {
        "operationId": "IdentityController_getTeam",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Team ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a team profile",
        "tags": [
          "identity"
        ]
      },
      "delete": {
        "operationId": "IdentityController_deleteTeam",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Team ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete a team",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/teams/{id}/capabilities": {
      "get": {
        "operationId": "IdentityController_getTeamCapabilities",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Team ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get aggregated team capabilities",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/teams/{id}/collaboration": {
      "get": {
        "operationId": "IdentityController_getTeamCollaboration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Team ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get collaboration pair scores for a team",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/task-completions": {
      "post": {
        "operationId": "IdentityController_createTaskCompletion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskCompletionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Record a task completion",
        "tags": [
          "identity"
        ]
      },
      "get": {
        "operationId": "IdentityController_queryTaskCompletions",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "taskId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "agentId",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Query task completions",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/delegation-templates": {
      "get": {
        "operationId": "IdentityController_getDelegationTemplates",
        "parameters": [
          {
            "name": "taskDescription",
            "required": true,
            "in": "query",
            "description": "Task to get suggestions for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get delegation template suggestions for a task",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/agents/{id}/trust-history": {
      "get": {
        "operationId": "IdentityController_getTrustHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agent ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": false,
            "in": "query",
            "description": "Number of days of history (default 30)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get trust score history for an agent over time",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/trust/bulk": {
      "get": {
        "operationId": "IdentityController_getBulkTrust",
        "parameters": [
          {
            "name": "agentIds",
            "required": true,
            "in": "query",
            "description": "Comma-separated agent IDs",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get trust profiles for multiple agents",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/agents/{id}/trust-profile": {
      "get": {
        "operationId": "IdentityController_getTrustProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agent ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get domain-specific trust profile for an agent",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/delegation-recall": {
      "get": {
        "operationId": "IdentityController_delegationRecall",
        "parameters": [
          {
            "name": "task",
            "required": true,
            "in": "query",
            "description": "Task description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get delegation-aware recall for a task",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/identity/agents/{id}/export": {
      "get": {
        "operationId": "IdentityController_exportIdentity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agent ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Export agent identity",
        "tags": [
          "identity"
        ]
      }
    },
...
  },
  "info": {
    "title": "Engram API",
    "description": "Memory infrastructure for AI agents",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AM-API-Key"
      },
      "user-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AM-User-ID"
      }
    },
    "schemas": {
...
      "CreateDelegationContractDto": {
        "type": "object",
        "properties": {
          "taskDescription": {
            "type": "string",
            "description": "Task description"
          },
          "expectedOutputs": {
            "description": "Expected outputs",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "successCriteria": {
            "description": "Success criteria",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timeout": {
            "type": "number",
            "description": "Timeout in ms"
          },
          "constraints": {
            "description": "Constraints",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "delegatedTo": {
            "type": "string",
            "description": "Agent ID to delegate to"
          },
          "accountId": {
            "type": "string",
            "description": "Account ID"
          }
        },
        "required": [
          "taskDescription",
          "expectedOutputs",
          "successCriteria",
          "timeout",
          "delegatedTo"
        ]
      },
...
      "CreateTeamDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}